refactor(ItemList, Item): Integrate Item into ItemList
