+++ src/components/ItemList.css
@@ -1,3 +1,5 @@
+/* CSS for ItemList */
+
 #best-items .item-list {
   margin-top: 16px;
   display: grid;
@@ -9,6 +11,47 @@
   display: grid;
   grid-template-columns: repeat(5, 1fr);
   gap: 40px 24px;
+}
+
+/* CSS for Item */
+
+.item-image {
+  width: 100%;
+  aspect-ratio: 1/1;
+  border-radius: 16px;
+}
+
+.item-info {
+  margin-top: 16px;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  text-align: left;
+  line-height: 1.2;
+  color: var(--cool-gray800);
+}
+
+.item-title {
+  font-size: 14px;
+  font-weight: 500;
+}
+
+.item-price {
+  font-size: 16px;
+  font-weight: 700;
+}
+
+.item-favorite {
+  display: flex;
+  font-size: 12px;
+  font-weight: 500;
+  align-items: center;
+}
+
+.item-favorite img {
+  width: 16px;
+  height: 16px;
+  margin-right: 4px;
 }
 
 @media (max-width: 1199px) {
